Refactor directory structure
Cleaned up the directory structure and file names in an effort to keep
some consistency.
