+++ client/src/Components/Authentication/Signup.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import { reduxForm } from 'redux-form';
-import * as actions from './AuthActions';
+import * as actions from './authAction';
 import axios from 'axios';
-import './index.css'; 
+import './auth.css'; 
 
 class Signup extends Component {
   constructor(props) {
