+++ client/src/Components/Dashboard/Dashboard.js
@@ -1,9 +1,8 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import * as actions from '../Authentication/AuthActions';
-import Header from '../App/Header';
-import Quiz from '../Quiz';
-import Meter from '../Meter';
+import Header from '../App/header';
+import Quiz from '../Quiz/quiz.js';
+import Meter from '../Meter/meter';
 
 export default class Dashboard extends Component {
