+++ client/src/components/LandingPage/landingPage.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import './index.css';
-import Header from '../App/Header';
+import './landingPage.css';
+import Header from '../App/header';
 
 function LandingPage() {
   return (
