+++ client/src/components/Quiz/quizAction.js
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { MARK_ANSWERED, FETCH_QUESTIONS } from '../constants/types';
+import { MARK_ANSWERED, FETCH_QUESTIONS } from '../../helpers/constants/types';
 import { browserHistory } from 'react-router';
 const apiUrl = 'http://localhost:3000/api/v1';
