+++ server/controllers/authentication.js
@@ -1,5 +1,5 @@
 const jwt = require('jwt-simple');
-const config = require('../../config');
+const config = require('../config');
 const Users = require(__dirname + '/../db/index').db.users;
 const Couples = require(__dirname + '/../db/index').db.couples;
 const CouplesUsers = require(__dirname + '/../db/index').db.couples_users;
